+++ backend/tests/routes/questions.test.js
@@ -1,8 +1,8 @@
 /* eslint-disable no-undef */
 // eslint-disable-next-line node/no-unpublished-require
 const request = require('supertest')
-const app = require('../app.js')
-const { initDatabase } = require('../config/setupDatabase')
+const app = require('../../app.js')
+const { initDatabase } = require('../../config/setupDatabase')
 
 beforeAll(async () => {
   await initDatabase()
