+++ backend/tests/routes/users.test.js
@@ -1,9 +1,9 @@
 /* eslint-disable no-undef */
 // eslint-disable-next-line node/no-unpublished-require
 const request = require('supertest')
-const app = require('../app.js')
-const { initDatabase } = require('../config/setupDatabase')
-const { User } = require('../models')
+const app = require('../../app.js')
+const { initDatabase } = require('../../config/setupDatabase')
+const { User } = require('../../models')
 
 beforeAll(async () => {
   await initDatabase()
